Updates to file structure and WorldAcousticSearch user interface
Moved database generator to be a method of main class. acoustics_database_catalog now only maintains catalog and region lists. Some helper functions moved to utilities.py
